Avoid busy loop on rsync exit status 12 or 13.
Although we set the next snapshot time to now + 60 seconds in case
rsync exits with exit status 12 or 13, we miss to check this time
barrier in case the snapshot creation status is HS_NEEDS_RESTART.

Fix this by adding an additional check in the switch() statement
of the select loop. As this change would trigger the assertion

	snapshot_creation_status == HS_READY

in next_snapshot_is_due(), remove this assertion.