Reuse old rsync argv if rsync has to be restarted.
If rsync must be restarted due to an exit code of 12 or 13,
create_rsync_argv() was called even if the old rsync_argv should
be reused in this case. This (correctly) triggers the assertion

	assert(!name_of_reference_snapshot);

in create_rsync_argv(). Fix this by not calling create_rsync_argv()
if there is a reference snapshot.