Don't create two snapshots in the same second.
This can only happen if all of the follwing are true:

	(a) source and destination directories are small
	(b) rsync completes successfully within one second
	(c) At most two snapshots are missing

In this case the rename() call which changes the snapshot name from
*-incomplete to the proper name fails for the second snapshot with
EEXIST. This is because the previous snapshot name coincides with
the name of the second snapshot.

The fix is a bit ugly but also non-invasive and simple: Just sleep
one second in this case.